+++ client/src/pages/PlacesFormPage.jsx
@@ -7,29 +7,7 @@ import api from "../utils/api";
 import { geocodeAddress } from "../utils/formUtils";
 import AddressSection from "../components/AddressSection";
 import AvailabilitySection from "../components/AvailabilitySection";
-
-// Helper function to validate and convert YouTube URL to embed format
-function extractYouTubeVideoId(url) {
-  if (!url || url.trim() === "") return "";
-  
-  // Regular expression patterns for different YouTube URL formats
-  const patterns = [
-    /(?:https?:\/\/)?(?:www\.)?youtube\.com\/watch\?v=([^&]+)/i,     // Regular URL format: youtube.com/watch?v=ID
-    /(?:https?:\/\/)?(?:www\.)?youtube\.com\/embed\/([^/?]+)/i,       // Embed URL format: youtube.com/embed/ID
-    /(?:https?:\/\/)?(?:www\.)?youtu\.be\/([^/?]+)/i,                 // Shortened URL format: youtu.be/ID
-    /(?:https?:\/\/)?(?:www\.)?youtube\.com\/shorts\/([^/?]+)/i       // Shorts URL format: youtube.com/shorts/ID
-  ];
-
-  for (const pattern of patterns) {
-    const match = url.match(pattern);
-    if (match && match[1]) {
-      // Return the URL in embed format
-      return `https://www.youtube.com/embed/${match[1]}`;
-    }
-  }
-
-  return "";  // Return empty string if invalid YouTube URL
-}
+import YouTubeSection, { extractYouTubeVideoId } from "../components/YouTubeSection";
 
 export default function PlacesFormPage() {
   const { id } = useParams();
@@ -477,13 +455,10 @@ export default function PlacesFormPage() {
           setAddedPhotos={setAddedPhotos}
         />
         
-        {preInput("YouTube Video", "add a YouTube link showcasing your conference room.")}
-        <input
-          type="text"
-          placeholder="https://www.youtube.com/watch?v=example"
-          value={youtubeLink}
-          onChange={(event) => setYoutubeLink(event.target.value)}
-          className="w-full border my-2 py-2 px-3 rounded-2xl"
+        <YouTubeSection
+          youtubeLink={youtubeLink}
+          setYoutubeLink={setYoutubeLink}
+          preInput={preInput}
         />
         
         {preInput("Description", "description of the conference room. ")}
